Implement multi-line input, with a serious caveat
Currently, we don't remove the newline that is in the multiline input.
This isn't the end of the world, but it isn't what bash does and could
therefore be surprising. Plus, it means your commands have to account
for it being there.

More critically though, we don't remove the slash itself, which means
this is effectively a useless feature until we remove it.